Add function to show thanks modal after form submission
This commit adds a new function called showThanksModal to display a modal dialog with a message after form submission. The function takes a message as a parameter and performs the following actions:

Hides the previous modal dialog
Opens a new modal dialog with the message
Adds the modal dialog to the DOM
After 4 seconds, removes the modal dialog and restores the previous modal dialog
Calls the closeModal function to close the modal dialog
The function uses DOM methods to create elements and add them to the DOM.

This commit improves the functionality of the code by providing a way to display a message to the user after form submission.